requirements.txt
aio-pika==9.6.0
aioboto3==15.5.0
aiobotocore==3.1.3
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aioitertools==0.13.0
aiormq==6.9.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asn1crypto==1.5.1
attrs==25.4.0
boto3==1.42.54
botocore==1.42.54
cachetools==7.0.1
cffi==2.0.0
click==8.3.1
cryptography==46.0.5
fastapi==0.129.2
frozenlist==1.8.0
greenlet==3.3.2
h11==0.16.0
idna==3.11
Jinja2==3.1.6
jmespath==1.1.0
jwcrypto==1.5.6
MarkupSafe==3.0.3
multidict==6.7.1
pamqp==3.3.0
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-pkcs11==0.9.3
PyYAML==6.0.3
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
wrapt==2.1.1
yarl==1.22.0
